Extract map and map layer selector from MapView.vue
